# docs/swagger.yaml
basePath: /api/v1
definitions:
  application.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  application.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  application.SubmitScoreRequest:
    description: Score submission request with game ID, score value, and optional
      metadata
    properties:
      game_id:
        example: game1
        type: string
      metadata:
        type: object
      score:
        example: 1000
        minimum: 0
        type: integer
    required:
    - game_id
    - score
    type: object
  response.ErrorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Real-time leaderboard system with WebSocket support for live updates
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Real-Time Leaderboard API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password, returns JWT access
        and refresh tokens
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using a valid refresh token. Returns new access
        and refresh tokens.
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password. Returns
        user information and JWT tokens.
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - auth
  /api/v1/leaderboard/game/{game_id}:
    get:
      consumes:
      - application/json
      description: Get paginated leaderboard for a specific game showing top players
      parameters:
      - description: Game ID
        example: game1
        in: path
        name: game_id
        required: true
        type: string
      - default: 10
        description: Number of top players
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Leaderboard retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get game leaderboard
      tags:
      - leaderboard
  /api/v1/leaderboard/global:
    get:
      consumes:
      - application/json
      description: Get paginated global leaderboard showing top players across all
        games
      parameters:
      - default: 10
        description: Number of top players
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Leaderboard retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get global leaderboard
      tags:
      - leaderboard
  /api/v1/leaderboard/rank/{user_id}:
    get:
      consumes:
      - application/json
      description: Get a user's rank, score, and position in a leaderboard (global
        or game-specific)
      parameters:
      - description: User ID
        example: user123
        in: path
        name: user_id
        required: true
        type: string
      - description: Game ID (empty or 'global' for global leaderboard)
        example: game1
        in: query
        name: game_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User rank retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found in leaderboard
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get user rank
      tags:
      - leaderboard
  /api/v1/reports/top-players:
    get:
      consumes:
      - application/json
      description: Get a paginated report of top players, optionally filtered by game
        and date range. Uses Redis for current data or PostgreSQL for historical data
        when date range is provided.
      parameters:
      - description: Game ID (empty or 'global' for global leaderboard)
        example: game1
        in: query
        name: game_id
        type: string
      - description: Start date in RFC3339 format
        example: "2024-01-01T00:00:00Z"
        in: query
        name: start_date
        type: string
      - description: End date in RFC3339 format
        example: "2024-01-31T23:59:59Z"
        in: query
        name: end_date
        type: string
      - default: 10
        description: Number of top players
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Report generated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get top players report
      tags:
      - reports
  /api/v1/scores:
    post:
      consumes:
      - application/json
      description: Submit a score for a game. Requires authentication. The score will
        be saved and the leaderboard will be updated.
      parameters:
      - description: Score submission request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.SubmitScoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Score submitted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Submit a score
      tags:
      - scores
  /api/v1/scores/me:
    get:
      consumes:
      - application/json
      description: Get paginated scores for the authenticated user. Optionally filter
        by game ID.
      parameters:
      - description: Filter by game ID
        example: game1
        in: query
        name: game_id
        type: string
      - default: 10
        description: Number of results per page
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Scores retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user scores
      tags:
      - scores
swagger: "2.0"